remove expire mechanism
The expiry mechanism didn't do what we want. If an
expiry interval wasn't set metrics would be sent out for an interval
even if no data was received, which is misleading, and if you set an
expiry interval then metrics would get marked inactive or disabled at
basically random times relative to when the metrics are sent,
potentially not sending metrics despite receiving data for them during a
reporting interval or sending metrics despite not receiving data for
them. Setting metric expiry values in brubeck_metric_find was obtuse,
since conceptually it's set when recording data for a metric.
Accidentally disabling metrics if the expiry timer is too short like in
[#33](https://github.com/github/brubeck/issues/33) shouldn't be a thing.
Disabling particular metrics via http calls seems unreliable due to
potential brubeck service restarts; it'd be better to have a pattern
blacklist in the config. Also, it wasn't threadsafe.

The behaviour we want is: if we've received data for a metric in some
interval, send it, and otherwise don't. To accomplish this, set metric
status as ACTIVE (ie send it at the end of this interval) when data is
recorded, and when the metric is sent set the status to INACTIVE so it
doesn't get sent again until a value for it is received. Some sort of
status/expiry mechanism is needed because memory is allocated for each
metric path only once and is kept forever.

See also [Dean's description of the
problem](https://github.com/github/brubeck/issues/51).