Fix iteration over multimsg batches
SIM_PACKETS is the max batch size for recvmmsg that was loaded from the
sampler's config, while the result from recvmmsg is the count of
messages that it received. Iterating past the count of received messages
will likely result in "packet_drop" messages in the log as we read
garbage memory.